% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner.R
\name{binner}
\alias{binner}
\title{Binning function for Time Series}
\usage{
binner(df, date_var, cat_var, val_var, interval = "day", sum_nm = "Total",
  count_nm = "Count", method = "sum")
}
\arguments{
\item{df}{Input df}

\item{date_var}{Unquoted name of your date variable (POSIX)}

\item{val_var}{Unquoted name of your value variable}

\item{sum_nm}{Char string of summing variable name, defauls to "Total"}

\item{count_nm}{Char string of count variable name, defauls to "Count"}

\item{method}{Defaults to "sum", or use "avg"/"mean"}
}
\description{
Get it a crack
}
