% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binner.R
\name{binner}
\alias{binner}
\title{Binning function for Time Series}
\usage{
binner(df, date_var, fact_var, val_var, interval = "day", sum_nm = "Total",
  count_nm = "Count")
}
\arguments{
\item{df}{Input df}

\item{date_var}{Unquoted name of your date variable (POSIX)}

\item{fact_var}{Unquoted name of your factor variable}

\item{val_var}{Unquoted name of your value variable}

\item{sum_nm}{Char string of summing variable name, defauls to "Total"}

\item{count_nm}{Char string of count variable name, defauls to "Count"}
}
\description{
Get it a crack
}
